:truck: merge did_key test modules (success tests with fail tests)
